FIX: repeat last line of .abc file
If the last line of the .abc file was empty and the second-to-last was
not, it could repeat the line fully or partially when reading the file,
leading to unexpected operations. This was due to how feof handles the
end of file, and was replaced with fgets in the loop condition.